fix: removido size_t

+++ src/matrix.c
@@ -7,7 +7,7 @@
 
 Vector Vector_read(FILE *fp) {
     Vector vec;
-    size_t capacity = 1000;
+    int capacity = 1000;
 
     vec.n = 0;
     vec.data = malloc(sizeof(float[capacity]));
@@ -33,10 +33,10 @@ Vector Vector_read(FILE *fp) {
         vec.data = realloc(vec.data, sizeof(float[vec.n]));
     }
 
-    vec.indices = malloc(sizeof(size_t[vec.n]));
+    vec.indices = malloc(sizeof(int[vec.n]));
     assert(vec.indices != NULL);
 
-    for(size_t i = 0; i < vec.n; i++) {
+    for(int i = 0; i < vec.n; i++) {
         vec.indices[i] = i;
     }
 
@@ -53,33 +53,33 @@ void Vector_free(Vector *vec) {
 
 
 void Vector_print(Vector *vec, FILE *fp) {
-    for(size_t i = 0; i < vec->n; i++) {
-        size_t pos = vec->indices[i];
+    for(int i = 0; i < vec->n; i++) {
+        int pos = vec->indices[i];
         fprintf(fp, "%f\n", vec->data[pos]);
     }
 }
 
 
-void Vector_swap(Vector *vec, size_t i, size_t j) {
+void Vector_swap(Vector *vec, int i, int j) {
     float f = vec->data[i];
     vec->data[i] = vec->data[j];
     vec->data[j] = f;
 
-    size_t s = vec->indices[i];
+    int s = vec->indices[i];
     vec->indices[i] = vec->indices[j];
     vec->indices[j] = s;
 }
 
 
-ColMajorMatrix Matrix_read(FILE *fp, size_t n) {
+ColMajorMatrix Matrix_read(FILE *fp, int n) {
     ColMajorMatrix mat;
 
     mat.n = n;
     mat.data = malloc(sizeof(float[n][n]));
     assert(mat.data != NULL);
 
-    for(size_t row = 0; row < n; row++) {
-        for(size_t col = 0; col < n; col++) {
+    for(int row = 0; row < n; row++) {
+        for(int col = 0; col < n; col++) {
             assert(!feof(fp));
             fscanf(fp, " %f", &mat.data[n*col + row]);
         }
@@ -96,8 +96,8 @@ void Matrix_free(ColMajorMatrix *mat) {
 
 
 void Matrix_print(ColMajorMatrix *mat, FILE *fp) {
-    for(size_t i = 0; i < mat->n; i++) {
-        for(size_t j = 0; j < mat->n; j++) {
+    for(int i = 0; i < mat->n; i++) {
+        for(int j = 0; j < mat->n; j++) {
             fprintf(fp, "%f ", mat->data[mat->n*j + i]);
         }
         fprintf(fp, "\n");
